Introduce loginAttempted() function
Call this function to indicate that you configured $http to provide an HTTP
Authentication header and trigger a retry of all deferred requests.
The first request will be retried: if it goes through, first
event:auth-loginSuccessful will be broadcast and then all other requests will
be retried. If it fails, event:auth-loginFailed will be broadcast and all
requests, including the first one, will be held in the buffer.
If any later request in the buffer fails, no special action will be taken, that
is, on 401 the whole process starts from the beginning, or on any other status
they just plain fail.