Fixed behaviour on failed login attempts
The request promise need to be removed from the retry queue and replaced
with a new promise in case it fails. To maintain order of the retry queue,
the request is marked with a config flag to be put in front of the retry queue
in case 401 is returned again.